team/requirements.txt
django
django-extensions 
djangorestframework 
requests
djangorestframework-simplejwt
django-cors-headers
jupyter
pandas
numpy
asgiref
mysqlclient
